refactor: organize pages into domain folders
