+++ MJ_FB_Frontend/src/App.tsx
@@ -1,21 +1,21 @@
 import { useState } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Profile from './components/Profile';
-import ManageAvailability from './components/StaffDashboard/ManageAvailability';
-import UserHistory from './components/StaffDashboard/UserHistory';
-import SlotBooking from './components/SlotBooking';
-import AddUser from './components/StaffDashboard/AddUser';
-import PantrySchedule from './components/StaffDashboard/PantrySchedule';
-import Pending from './components/StaffDashboard/Pending';
-import Login from './components/Login';
-import StaffLogin from './components/StaffLogin';
-import VolunteerLogin from './components/VolunteerLogin';
-import VolunteerDashboard from './components/VolunteerDashboard';
-import VolunteerManagement from './components/VolunteerManagement';
+import Profile from './pages/booking/Profile';
+import ManageAvailability from './pages/staff/ManageAvailability';
+import UserHistory from './pages/staff/UserHistory';
+import SlotBooking from './pages/booking/SlotBooking';
+import AddUser from './pages/staff/AddUser';
+import PantrySchedule from './pages/staff/PantrySchedule';
+import Pending from './pages/staff/Pending';
+import Login from './pages/auth/Login';
+import StaffLogin from './pages/auth/StaffLogin';
+import VolunteerLogin from './pages/auth/VolunteerLogin';
+import VolunteerDashboard from './pages/volunteer/VolunteerDashboard';
+import VolunteerManagement from './pages/volunteer-management/VolunteerManagement';
 import Dashboard from './pages/Dashboard';
 import UserDashboard from './pages/UserDashboard';
-import VolunteerBookingHistory from './components/VolunteerBookingHistory';
-import VolunteerSchedule from './components/VolunteerSchedule';
+import VolunteerBookingHistory from './pages/volunteer/VolunteerBookingHistory';
+import VolunteerSchedule from './pages/volunteer/VolunteerSchedule';
 import WarehouseDashboard from './pages/WarehouseDashboard';
 import DonationLog from './pages/DonationLog';
 import TrackPigpound from './pages/TrackPigpound';
@@ -26,7 +26,7 @@ import DonorProfile from './pages/DonorProfile';
 import AdminStaffList from './pages/AdminStaffList';
 import AdminStaffForm from './pages/AdminStaffForm';
 import Events from './pages/Events';
-import PantryVisits from './components/StaffDashboard/PantryVisits';
+import PantryVisits from './pages/staff/PantryVisits';
 import Navbar, { type NavGroup, type NavLink } from './components/Navbar';
 import FeedbackSnackbar from './components/FeedbackSnackbar';
 import Breadcrumbs from './components/Breadcrumbs';
