+++ MJ_FB_Frontend/src/pages/AdminStaffForm.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import StaffForm from '../components/StaffDashboard/StaffForm';
+import StaffForm from '../components/StaffForm';
 import { getStaff, createStaff, updateStaff } from '../api/adminStaff';
 import FeedbackSnackbar from '../components/FeedbackSnackbar';
 import type { Staff } from '../types';
