+++ MJ_FB_Frontend/src/pages/staff/Pending.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Grid, Card, CardContent, CardActions, Typography, TextField, Button } from '@mui/material';
 import { getBookings, decideBooking } from '../../api/bookings';
-import FeedbackSnackbar from '../FeedbackSnackbar';
+import FeedbackSnackbar from '../../components/FeedbackSnackbar';
 import { formatTime } from '../../utils/time';
 
 interface Booking {
